crpyto_price_chainlit/pyproject.toml
[project]
name = "crpyto-price-chainlit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.5.5",
    "openai-agents>=0.0.19",
    "python-dotenv>=1.1.1",
]
